+++ src/daf/message/text_based.py
@@ -38,13 +38,9 @@ class TextMESSAGE(BaseMESSAGE):
     """
     This class is used for creating objects that represent messages which will be sent to Discord's TEXT CHANNELS.
 
-    .. versionchanged:: v2.3
+    .. versionchanged:: v2.7
 
-        :Slow mode period handling:
-
-            When the period is lower than the remaining time, the framework will start
-            incrementing the original period by original period until it is larger then
-            the slow mode remaining time.
+        *start_in* now accepts datetime object
 
     Parameters
     ------------
@@ -100,8 +96,9 @@ class TextMESSAGE(BaseMESSAGE):
         - "send" -    each period a new message will be sent,
         - "edit" -    each period the previously send message will be edited (if it exists)
         - "clear-send" -    previous message will be deleted and a new one sent.
-    start_in: Optional[timedelta]
+    start_in: Optional[timedelta | datetime]
         When should the message be first sent.
+        *timedelta* means the difference from current time, while *datetime* means actual first send time.
     remove_after: Optional[Union[int, timedelta, datetime]]
         Deletes the message after:
 
@@ -123,7 +120,7 @@ class TextMESSAGE(BaseMESSAGE):
                  data: Union[Iterable[Union[str, discord.Embed, FILE]], str, discord.Embed, FILE, _FunctionBaseCLASS],
                  channels: Union[Iterable[Union[int, discord.TextChannel, discord.Thread]], AutoCHANNEL],
                  mode: Optional[Literal["send", "edit", "clear-send"]] = "send",
-                 start_in: Optional[Union[timedelta, bool]] = timedelta(seconds=0),
+                 start_in: Optional[Union[timedelta, datetime]] = datetime.now(),
                  remove_after: Optional[Union[int, timedelta, datetime]] = None):
         super().__init__(start_period, end_period, data, start_in, remove_after)
         self.mode = mode
@@ -317,6 +314,9 @@ class TextMESSAGE(BaseMESSAGE):
         ValueError
             No valid channels were passed to object"
         """
+        if parent is None:
+            return
+
         ch_i = 0
         self.parent = parent
         cl = parent.parent.client
@@ -544,7 +544,7 @@ class TextMESSAGE(BaseMESSAGE):
         """
         if "start_in" not in kwargs:
             # This parameter does not appear as attribute, manual setting necessary
-            kwargs["start_in"] = timedelta(seconds=0)
+            kwargs["start_in"] = self.next_send_time
 
         if "data" not in kwargs:
             kwargs["data"] = self._data
@@ -569,14 +569,11 @@ class DirectMESSAGE(BaseMESSAGE):
 
     .. deprecated:: v2.1
 
-        - start_in (start_now) - Using bool value to dictate whether the message should be sent at framework start.
         - start_period, end_period - Using int values, use ``timedelta`` object instead.
 
-    .. versionchanged:: v2.1
+    .. versionchanged:: v2.7
 
-        - start_period, end_period Accept timedelta objects.
-        - start_now - renamed into ``start_in`` which describes when the message should be first sent.
-        - removed ``deleted`` property
+        *start_in* now accepts datetime object
 
     Parameters
     ------------
@@ -630,8 +627,9 @@ class DirectMESSAGE(BaseMESSAGE):
         - "edit" -    each period the previously send message will be edited (if it exists)
         - "clear-send" -    previous message will be deleted and a new one sent.
 
-    start_in: Optional[timedelta]
+    start_in: Optional[timedelta | datetime]
         When should the message be first sent.
+        *timedelta* means the difference from current time, while *datetime* means actual first send time.
     remove_after: Optional[Union[int, timedelta, datetime]]
         Deletes the guild after:
 
@@ -652,7 +650,7 @@ class DirectMESSAGE(BaseMESSAGE):
                  end_period: Union[int, timedelta],
                  data: Union[str, discord.Embed, FILE, Iterable[Union[str, discord.Embed, FILE]], _FunctionBaseCLASS],
                  mode: Optional[Literal["send", "edit", "clear-send"]] = "send",
-                 start_in: Optional[Union[timedelta, bool]] = timedelta(seconds=0),
+                 start_in: Optional[Union[timedelta, datetime]] = datetime.now(),
                  remove_after: Optional[Union[int, timedelta, datetime]] = None):
         super().__init__(start_period, end_period, data, start_in, remove_after)
         self.mode = mode
@@ -768,6 +766,9 @@ class DirectMESSAGE(BaseMESSAGE):
             Raised when the direct message channel could not be created
         """
         try:
+            if parent is None:
+                return
+
             self.parent = parent
             user = parent.apiobject
             await user.create_dm()
@@ -892,10 +893,9 @@ class DirectMESSAGE(BaseMESSAGE):
         Other
             Raised from .initialize() method
         """
-
         if "start_in" not in kwargs:
             # This parameter does not appear as attribute, manual setting necessary
-            kwargs["start_in"] = timedelta(seconds=0)
+            kwargs["start_in"] = self.next_send_time
 
         if "data" not in kwargs:
             kwargs["data"] = self._data