+++ src/daf_gui/main.py
@@ -19,10 +19,9 @@ import webbrowser
 
 try:
     from .widgets import *
-    from .convert import *
 except ImportError:
     from widgets import *
-    from convert import *
+
 
 
 WIN_UPDATE_DELAY = 0.005
@@ -39,10 +38,49 @@ GITHUB_URL = "https://github.com/davidhozic/discord-advertisement-framework"
 DOC_URL = f"https://daf.davidhozic.com/en/v{daf.VERSION}"
 
 
+class GLOBAL:
+    app: "Application" = None
+
+
+def gui_except(fnc: Callable):
+    """
+    Decorator that catches exceptions and displays them in GUI.
+    """
+    def wrapper(*args, **kwargs):
+        try:
+            return fnc(*args, **kwargs)
+        except Exception as exc:
+            tkdiag.Messagebox.show_error(f"{exc}\n(Exception in {fnc.__name__})")              
+
+    return wrapper
+
+
+def gui_confirm_action(fnc: Callable):
+    """
+    Decorator that asks the user to confirm the action before calling the
+    targeted function (fnc).
+    """
+    def wrapper(*args, **kwargs):
+        result = tkdiag.Messagebox.show_question("Are you sure?", "Confirm")
+        if result == "Yes":
+            return fnc(*args, **kwargs)
+
+    return wrapper
+
+
+def gui_daf_assert_running():
+    if not GLOBAL.app._daf_running:
+        raise ConnectionError("Start the framework first (START button)")
+
+
 class Application():
     def __init__(self) -> None:
         # Window initialization
         win_main = ttk.Window(themename="cosmo")
+
+        # DPI
+        set_dpi(win_main.winfo_fpixels('1i'))
+        dpi_5 = dpi_scaled(5)
         # path = os.path.join(os.path.dirname(__file__), "img/logo.png")
         # photo = tk.PhotoImage(file=path)
         # win_main.iconphoto(True, photo)
@@ -53,12 +91,9 @@ class Application():
         win_main.wm_minsize(*screen_res)
         win_main.protocol("WM_DELETE_WINDOW", self.close_window)
 
-        # Console initialization
-        self.win_debug = None
-
         # Toolbar
         self.frame_toolbar = ttk.Frame(self.win_main)
-        self.frame_toolbar.pack(fill=tk.X, side="top", padx=5, pady=5)
+        self.frame_toolbar.pack(fill=tk.X, side="top", padx=dpi_5, pady=dpi_5)
         self.bnt_toolbar_start_daf = ttk.Button(self.frame_toolbar, text="Start", command=self.start_daf)
         self.bnt_toolbar_start_daf.pack(side="left")
         self.bnt_toolbar_stop_daf = ttk.Button(self.frame_toolbar, text="Stop", state="disabled", command=self.stop_daf)
@@ -74,6 +109,9 @@ class Application():
         # Objects tab
         self.init_schema_tab()
 
+        # Live inspect tab
+        self.init_live_inspect_tab()
+
         # Output tab
         self.init_output_tab()
 
@@ -87,20 +125,19 @@ class Application():
         self._daf_running = False
         self._window_opened = True
 
-        # Tasks
-        self._async_queue = asyncio.Queue()
-
         # On close configuration
         self.win_main.protocol("WM_DELETE_WINDOW", self.close_window)
 
     def init_schema_tab(self):
         self.objects_edit_window = None
+        dpi_10 = dpi_scaled(10)
+        dpi_5 = dpi_scaled(5)
 
-        tab_schema = ttk.Frame(self.tabman_mf, padding=(10, 10))
+        tab_schema = ttk.Frame(self.tabman_mf, padding=(dpi_10, dpi_10))
         self.tabman_mf.add(tab_schema, text="Schema definition")
 
         # Object tab file menu
-        bnt_file_menu = ttk.Menubutton(tab_schema, text="Load/Save/Generate")
+        bnt_file_menu = ttk.Menubutton(tab_schema, text="Schema")
         menubar_file = ttk.Menu(bnt_file_menu)
         menubar_file.add_command(label="Save schema", command=self.save_schema)
         menubar_file.add_command(label="Load schema", command=self.load_schema)
@@ -109,30 +146,62 @@ class Application():
         bnt_file_menu.pack(anchor=tk.W)
 
         # Object tab account tab
-        frame_tab_account = ttk.Labelframe(tab_schema, text="Accounts", padding=(10, 10), bootstyle="primary")
-        frame_tab_account.pack(side="left", fill=tk.BOTH, expand=True, pady=10, padx=5)
+        frame_tab_account = ttk.Labelframe(
+            tab_schema,
+            text="Accounts", padding=(dpi_10, dpi_10), bootstyle="primary")
+        frame_tab_account.pack(side="left", fill=tk.BOTH, expand=True, pady=dpi_10, padx=dpi_5)
 
-        frame_account_bnts = ttk.Frame(frame_tab_account, padding=(0, 10))
-        frame_account_bnts.pack(fill=tk.X)
-        self.bnt_add_object = ttk.Button(
-            frame_account_bnts,
-            text="Add ACCOUNT",
+        @gui_except
+        @gui_confirm_action
+        def import_accounts():
+            "Imports account from live view"
+            accs = daf.get_accounts()
+            for acc in accs:
+                acc.intents = None  # Intents cannot be loaded properly
+
+            values = convert_to_object_info(accs, save_original=False)
+            if not len(values):
+                raise ValueError("Live view has no elements.")
+
+            self.lb_accounts.clear()
+            self.lb_accounts.insert(tk.END, *values)
+
+        menu_bnt = ttk.Menubutton(
+            frame_tab_account,
+            text="Object options"
+        )
+        menu = ttk.Menu(menu_bnt)
+        menu.add_command(
+            label="New ACCOUNT",
             command=lambda: self.open_object_edit_window(daf.ACCOUNT, self.lb_accounts)
         )
-        self.bnt_edit_object = ttk.Button(frame_account_bnts, text="Edit", command=self.edit_accounts)
-        self.bnt_remove_object = ttk.Button(frame_account_bnts, text="Remove", command=self.list_del_account)
-        self.bnt_add_object.pack(side="left")
-        self.bnt_edit_object.pack(side="left")
-        self.bnt_remove_object.pack(side="left")
+        menu.add_command(label="Edit", command=self.edit_accounts)
+        menu.add_command(label="Remove", command=self.list_del_account)
+        menu_bnt.configure(menu=menu)
+        menu_bnt.pack(anchor=tk.W)
 
-        self.lb_accounts = ListBoxScrolled(frame_tab_account, background="#000")
-        self.lb_accounts.pack(fill=tk.BOTH, expand=True)
+        frame_account_bnts = ttk.Frame(frame_tab_account)
+        frame_account_bnts.pack(fill=tk.X, pady=dpi_5)
+        ttk.Button(
+            frame_account_bnts, text="Import from live", command=import_accounts
+        ).pack(side="left")
+        ttk.Button(
+            frame_account_bnts, text="Load selection to live", command=lambda: self.add_accounts_daf(True)
+        ).pack(side="left")
+        self.load_at_start_var = ttk.BooleanVar(value=True)
+        ttk.Checkbutton(
+            frame_account_bnts, text="Load all at start", onvalue=True, offvalue=False, state="normal",
+            variable=self.load_at_start_var
+        ).pack(side="left", padx=dpi_5)
+
+        self.lb_accounts = ListBoxScrolled(frame_tab_account)
+        self.lb_accounts.pack(fill=tk.BOTH, expand=True, side="left")
 
         # Object tab account tab logging tab
-        frame_logging = ttk.Labelframe(tab_schema, padding=(10, 10), text="Logging", bootstyle="primary")
+        frame_logging = ttk.Labelframe(tab_schema, padding=(dpi_10, dpi_10), text="Logging", bootstyle="primary")
         label_logging_mgr = ttk.Label(frame_logging, text="Selected logger:")
         label_logging_mgr.pack(anchor=tk.N)
-        frame_logging.pack(side="left", fill=tk.BOTH, expand=True, pady=10, padx=5)
+        frame_logging.pack(side="left", fill=tk.BOTH, expand=True, pady=dpi_10, padx=dpi_5)
 
         frame_logger_select = ttk.Frame(frame_logging)
         frame_logger_select.pack(fill=tk.X)
@@ -155,6 +224,75 @@ class Application():
         ]
 
         self.combo_tracing["values"] = [en for en in daf.TraceLEVELS]
+
+    def init_live_inspect_tab(self):
+        dpi_10 = dpi_scaled(10)
+        dpi_5 = dpi_scaled(5)
+
+        def remove_account():
+            selection = list_live_objects.curselection()
+            if len(selection):
+                @gui_confirm_action
+                def _():
+                    values = list_live_objects.get()
+                    for i in selection:
+                        async_execute(
+                            daf.remove_object(values[i].real_object),
+                            parent_window=self.win_main
+                        )
+
+                    async_execute(dummy_task(), lambda x: self.load_live_accounts(), self.win_main)
+
+                _()
+            else:
+                tkdiag.Messagebox.show_error("Select atlest one item!", "Select errror")
+
+        @gui_except
+        def add_account():
+            gui_daf_assert_running()
+            selection = combo_add_object_edit.combo.current()
+            if selection >= 0:
+                fnc: ObjectInfo = combo_add_object_edit.combo.get()
+                mapping = {k: convert_to_objects(v) for k, v in fnc.data.items()}
+                async_execute(fnc.class_(**mapping), parent_window=self.win_main)
+            else:
+                tkdiag.Messagebox.show_error("Combobox does not have valid selection.", "Combo invalid selection")
+
+        def view_live_account():
+            selection = list_live_objects.curselection()
+            if len(selection) == 1:
+                object_: ObjectInfo = list_live_objects.get()[selection[0]]
+                self.open_object_edit_window(
+                    daf.ACCOUNT,
+                    list_live_objects,
+                    old=object_
+                )
+            else:
+                tkdiag.Messagebox.show_error("Select one item!", "Empty list!")
+
+        tab_live = ttk.Frame(self.tabman_mf, padding=(dpi_10, dpi_10))
+        self.tabman_mf.add(tab_live, text="Live view")
+        frame_add_account = ttk.Frame(tab_live)
+        frame_add_account.pack(fill=tk.X, pady=dpi_10)
+
+        combo_add_object_edit = ComboEditFrame(
+            self,
+            [ObjectInfo(daf.add_object, {})],
+            master=frame_add_account,
+            check_parameters=False
+        )
+        ttk.Button(frame_add_account, text="Execute", command=add_account).pack(side="left")
+        combo_add_object_edit.pack(side="left", fill=tk.X, expand=True)
+
+        frame_account_opts = ttk.Frame(tab_live)
+        frame_account_opts.pack(fill=tk.X, pady=dpi_5)
+        ttk.Button(frame_account_opts, text="Refresh", command=self.load_live_accounts).pack(side="left")
+        ttk.Button(frame_account_opts, text="Edit", command=view_live_account).pack(side="left")
+        ttk.Button(frame_account_opts, text="Remove", command=remove_account).pack(side="left")
+
+        list_live_objects = ListBoxScrolled(tab_live)
+        list_live_objects.pack(fill=tk.BOTH, expand=True)
+        self.list_live_objects = list_live_objects
 
     def init_output_tab(self):
         self.tab_output = ttk.Frame(self.tabman_mf)
@@ -186,6 +324,8 @@ class Application():
         sys.stdout = STDIOOutput()
 
     def init_credits_tab(self):
+        dpi_10 = dpi_scaled(10)
+        dpi_30 = dpi_scaled(30)
         logo_img = Image.open(f"{os.path.dirname(__file__)}/img/logo.png")
         logo_img = logo_img.resize(
             (self.win_main.winfo_screenwidth() // 8, self.win_main.winfo_screenwidth() // 8),
@@ -195,41 +335,44 @@ class Application():
         self.tab_info = ttk.Frame(self.tabman_mf)
         self.tabman_mf.add(self.tab_info, text="About")
         info_bnts_frame = ttk.Frame(self.tab_info)
-        info_bnts_frame.pack(pady=30)
+        info_bnts_frame.pack(pady=dpi_30)
         ttk.Button(info_bnts_frame, text="Github", command=lambda: webbrowser.open(GITHUB_URL)).grid(row=0, column=0)
         ttk.Button(
             info_bnts_frame,
             text="Documentation",
             command=lambda: webbrowser.open(DOC_URL)
         ).grid(row=0, column=1)
-        ttk.Label(self.tab_info, text="Like the app? Give it a star :) on GitHub (^)").pack(pady=10)
+        ttk.Label(self.tab_info, text="Like the app? Give it a star :) on GitHub (^)").pack(pady=dpi_10)
         ttk.Label(self.tab_info, text=CREDITS_TEXT).pack()
         label_logo = ttk.Label(self.tab_info, image=logo)
         label_logo.image = logo
         label_logo.pack()
 
     def init_analytics_tab(self):
-        tab_analytics = ttk.Frame(self.tabman_mf, padding=(10, 10))
+        dpi_10 = dpi_scaled(10)
+        dpi_5 = dpi_scaled(5)
+        tab_analytics = ttk.Frame(self.tabman_mf, padding=(dpi_10, dpi_10))
         self.tabman_mf.add(tab_analytics, text="Analytics")
 
         # Message log
-        frame_msg_history = ttk.Labelframe(tab_analytics, padding=(10, 10), text="Messages", bootstyle="primary")
+        frame_msg_history = ttk.Labelframe(tab_analytics, padding=(dpi_10, dpi_10), text="Messages", bootstyle="primary")
         frame_msg_history.pack(fill=tk.BOTH, expand=True)
 
         frame_combo_messages = ComboEditFrame(
-            self.edit_analytics_messages,
-            ObjectInfo(daf.logging.LoggerBASE.analytic_get_message_log, {}),
-            frame_msg_history
+            self,
+            [ObjectInfo(daf.logging.LoggerBASE.analytic_get_message_log, {})],
+            frame_msg_history,
+            check_parameters=False
         )
         frame_combo_messages.pack(fill=tk.X)
         self.frame_combo_messages = frame_combo_messages
 
         frame_msg_history_bnts = ttk.Frame(frame_msg_history)
-        frame_msg_history_bnts.pack(fill=tk.X, pady=10)
+        frame_msg_history_bnts.pack(fill=tk.X, pady=dpi_10)
         ttk.Button(
             frame_msg_history_bnts,
             text="Get logs",
-            command=lambda: self._async_queue.put_nowait(self.analytics_load_msg())
+            command=lambda: async_execute(self.analytics_load_msg(), parent_window=self.win_main)
         ).pack(side="left", fill=tk.X)
         ttk.Button(frame_msg_history_bnts, command=self.show_message_log, text="View log").pack(side="left", fill=tk.X)
         ttk.Button(
@@ -244,11 +387,12 @@ class Application():
         self.lst_message_log = lst_messages
 
         # Number of messages
-        frame_num_msg = ttk.Labelframe(tab_analytics, padding=(10, 10), text="Number of messages", bootstyle="primary")
+        frame_num_msg = ttk.Labelframe(tab_analytics, padding=(dpi_10, dpi_10), text="Number of messages", bootstyle="primary")
         frame_combo_num_messages = ComboEditFrame(
-            self.edit_analytics_num_msg,
-            ObjectInfo(daf.logging.LoggerBASE.analytic_get_num_messages, {}),
-            frame_num_msg
+            self,
+            [ObjectInfo(daf.logging.LoggerBASE.analytic_get_num_messages, {})],
+            frame_num_msg,
+            check_parameters=False
         )
 
         coldata = [
@@ -272,10 +416,10 @@ class Application():
         ttk.Button(
             frame_num_msg,
             text="Calculate",
-            command=lambda: self._async_queue.put_nowait(self.analytics_load_num_msg())
-        ).pack(anchor=tk.W, pady=10)
+            command=lambda: async_execute(self.analytics_load_num_msg(), parent_window=self.win_main)
+        ).pack(anchor=tk.W, pady=dpi_10)
 
-        frame_num_msg.pack(fill=tk.BOTH, expand=True, pady=5)
+        frame_num_msg.pack(fill=tk.BOTH, expand=True, pady=dpi_5)
         tw_num_msg.pack(expand=True, fill=tk.BOTH)
 
         self.tw_num_msg = tw_num_msg
@@ -290,7 +434,13 @@ class Application():
             self.objects_edit_window = ObjectEditWindow()
             self.objects_edit_window.open_object_edit_frame(*args, **kwargs)
 
+    def load_live_accounts(self):
+        object_infos = convert_to_object_info(daf.get_accounts(), save_original=True)
+        self.list_live_objects.clear()
+        self.list_live_objects.insert(tk.END, *object_infos)
+
     async def analytics_load_msg(self):
+        gui_daf_assert_running()
         logger = daf.get_logger()
         if not isinstance(logger, daf.LoggerSQL):
             raise ValueError("Analytics only allowed when using LoggerSQL")
@@ -304,11 +454,12 @@ class Application():
         messages = await logger.analytic_get_message_log(
             **data
         )
-        messages = convert_to_object_info(messages)
+        messages = convert_to_object_info(messages, cache=True)
         self.lst_message_log.clear()
         self.lst_message_log.insert(tk.END, *messages)
 
     async def analytics_load_num_msg(self):
+        gui_daf_assert_running()
         logger = daf.get_logger()
         if not isinstance(logger, daf.LoggerSQL):
             raise ValueError("Analytics only allowed when using LoggerSQL")
@@ -357,32 +508,6 @@ class Application():
         else:
             tkdiag.Messagebox.show_error("Select ONE item!", "Empty list!")
 
-    def edit_analytics_messages(self):
-        selection = self.frame_combo_messages.combo.current()
-        if selection >= 0:
-            object_: ObjectInfo = self.frame_combo_messages.combo.get()
-            self.open_object_edit_window(
-                daf.logging.LoggerBASE.analytic_get_message_log,
-                self.frame_combo_messages.combo,
-                old=object_,
-                check_parameters=False
-            )
-        else:
-            tkdiag.Messagebox.show_error("Select atleast one item!", "Empty list!")
-
-    def edit_analytics_num_msg(self):
-        selection = self.frame_combo_num_messages.combo.current()
-        if selection >= 0:
-            object_: ObjectInfo = self.frame_combo_num_messages.combo.get()
-            self.open_object_edit_window(
-                daf.logging.LoggerBASE.analytic_get_num_messages,
-                self.frame_combo_num_messages.combo,
-                old=object_,
-                check_parameters=False
-            )
-        else:
-            tkdiag.Messagebox.show_error("Select atleast one item!", "Empty list!")
-
     def edit_logger(self):
         selection = self.combo_logging_mgr.current()
         if selection >= 0:
@@ -402,7 +527,10 @@ class Application():
     def list_del_account(self):
         selection = self.lb_accounts.curselection()
         if len(selection):
-            self.lb_accounts.delete(*selection)
+            @gui_confirm_action
+            def _():
+                self.lb_accounts.delete(*selection)
+            _()
         else:
             tkdiag.Messagebox.show_error("Select atleast one item!", "Empty list!")
 
@@ -468,6 +596,7 @@ daf.run(
         if not file.name.endswith(".py"):
             os.rename(file.name, file.name + ".py")
 
+    @gui_except
     def save_schema(self) -> bool:
         filename = tkfile.asksaveasfilename(filetypes=[("JSON", "*.json")])
         if filename == "":
@@ -475,78 +604,89 @@ daf.run(
 
         json_data = {
             "loggers": {
-                "all": [convert_to_json(x) for x in self.combo_logging_mgr["values"]],
+                "all": convert_to_json(self.combo_logging_mgr["values"]),
                 "selected_index": self.combo_logging_mgr.current(),
             },
             "tracing": self.combo_tracing.current(),
-            "accounts": [convert_to_json(x) for x in self.lb_accounts.get()],
+            "accounts": convert_to_json(self.lb_accounts.get()),
         }
+
+        if not filename.endswith(".json"):
+            filename += ".json"
 
         with open(filename, "w", encoding="utf-8") as file:
             json.dump(json_data, file, indent=2)
 
-        if not filename.endswith(".json"):
-            os.rename(filename, filename + ".json")
+        tkdiag.Messagebox.show_info(f"Saved to {filename}", "Finished", self.win_main)
 
         return True
 
+    @gui_except
     def load_schema(self):
-        try:
-            filename = tkfile.askopenfilename(filetypes=[("JSON", "*.json")])
-            if filename == "":
-                return
+        filename = tkfile.askopenfilename(filetypes=[("JSON", "*.json")])
+        if filename == "":
+            return
 
-            with open(filename, "r", encoding="utf-8") as file:
-                json_data = json.load(file)
+        with open(filename, "r", encoding="utf-8") as file:
+            json_data = json.load(file)
 
-                # Load accounts
-                accounts = convert_from_json(json_data["accounts"])
-                self.lb_accounts.clear()
-                self.lb_accounts.listbox.insert(tk.END, *accounts)
+            # Load accounts
+            accounts = convert_from_json(json_data["accounts"])
+            self.lb_accounts.clear()
+            self.lb_accounts.listbox.insert(tk.END, *accounts)
 
-                # Load loggers
-                loggers = [convert_from_json(x) for x in json_data["loggers"]["all"]]
-                self.combo_logging_mgr["values"] = loggers
-                selected_index = json_data["loggers"]["selected_index"]
-                if selected_index >= 0:
-                    self.combo_logging_mgr.current(selected_index)
+            # Load loggers
+            loggers = [convert_from_json(x) for x in json_data["loggers"]["all"]]
+            self.combo_logging_mgr["values"] = loggers
+            selected_index = json_data["loggers"]["selected_index"]
+            if selected_index >= 0:
+                self.combo_logging_mgr.current(selected_index)
 
-                # Tracing
-                tracing_index = json_data["tracing"]
-                if tracing_index >= 0:
-                    self.combo_tracing.current(json_data["tracing"])
+            # Tracing
+            tracing_index = json_data["tracing"]
+            if tracing_index >= 0:
+                self.combo_tracing.current(json_data["tracing"])
 
-        except Exception as exc:
-            tkdiag.Messagebox.show_error(f"Could not load schema!\n\n{exc}", "Schema load error!")
-
+    @gui_except
     def start_daf(self):
-        try:
-            logger = self.combo_logging_mgr.get()
-            if isinstance(logger, str) and logger == "":
-                logger = None
-            elif logger is not None:
-                logger = convert_to_objects(logger)
+        logger = self.combo_logging_mgr.get()
+        if isinstance(logger, str) and logger == "":
+            logger = None
+        elif logger is not None:
+            logger = convert_to_objects(logger)
 
-            tracing = self.combo_tracing.get()
-            if isinstance(tracing, str) and tracing == "":
-                tracing = None
+        tracing = self.combo_tracing.get()
+        if isinstance(tracing, str) and tracing == "":
+            tracing = None
 
-            self._async_queue.put_nowait(daf.initialize(logger=logger, debug=tracing))
-            self._async_queue.put_nowait(
-                [daf.add_object(convert_to_objects(account)) for account in self.lb_accounts.get()]
-            )
-            self.bnt_toolbar_start_daf.configure(state="disabled")
-            self.bnt_toolbar_stop_daf.configure(state="enabled")
-            self._daf_running = True
-        except Exception as exc:
-            print(exc)
-            tkdiag.Messagebox.show_error(f"Could not start daf due to exception!\n\n{exc}", "Start error!")
+        async_execute(daf.initialize(logger=logger, debug=tracing), parent_window=self.win_main)
+        self._daf_running = True
+        if self.load_at_start_var.get():
+            self.add_accounts_daf()
+
+        self.bnt_toolbar_start_daf.configure(state="disabled")
+        self.bnt_toolbar_stop_daf.configure(state="enabled")
 
     def stop_daf(self):
-        self._async_queue.put_nowait(daf.shutdown())
+        async_execute(daf.shutdown(), parent_window=self.win_main)
         self._daf_running = False
         self.bnt_toolbar_start_daf.configure(state="enabled")
         self.bnt_toolbar_stop_daf.configure(state="disabled")
+
+    @gui_except
+    def add_accounts_daf(self, selection: bool = False):
+        gui_daf_assert_running()
+        accounts = self.lb_accounts.get()
+        if selection:
+            indexes = self.lb_accounts.curselection()
+            if not len(indexes):
+                raise ValueError("Select at least one item.")
+
+            indexes = set(indexes)
+            accounts = [a for i, a in enumerate(accounts) if i in indexes]
+
+        for account in accounts:
+            async_execute(daf.add_object(convert_to_objects(account)), parent_window=self.win_main)
 
     def close_window(self):
         resp = tkdiag.Messagebox.yesnocancel("Do you wish to save?", "Save?", alert=True, parent=self.win_main)
@@ -562,39 +702,25 @@ daf.run(
             self.win_main.destroy()
             self.win_main.quit()
 
-        self._async_queue.put_nowait(_tmp())
-
-    async def _run_coro_gui_errors(self, coro: Awaitable):
-        try:
-            await coro
-        except asyncio.QueueEmpty:
-            raise
-        except Exception as exc:
-            tkdiag.Messagebox.show_error(str(exc), "Coroutine error")
+        async_execute(_tmp())
 
     async def _process(self):
         self.win_main.update()
-        try:
-            t = self._async_queue.get_nowait()
-            if isinstance(t, Iterable):
-                for c in t:
-                    asyncio.create_task(self._run_coro_gui_errors(c))
-            else:
-                await self._run_coro_gui_errors(t)
-        except asyncio.QueueEmpty:
-            pass
 
 
 def run():
-    win_main = Application()
+    app = Application()
+    GLOBAL.app = app
 
     async def update_task():
-        while win_main.opened:
-            await win_main._process()
+        while app.opened:
+            await app._process()
             await asyncio.sleep(WIN_UPDATE_DELAY)
 
     loop = asyncio.new_event_loop()
+    async_start(loop)
     loop.run_until_complete(update_task())
+    loop.run_until_complete(async_stop())
 
 
 if __name__ == "__main__":